Reference

Sound Construction

Sounds.Sound — Type.

Sound(x::AbstractArray;[rate=samplerate(x)])

Creates a sound object from an array.

Assumes 1 is the loudest and -1 the softest. The array should be 1d for mono signals, or an array of size (N,2) for stereo sounds. If samplerate is defined for this type of array, it will be used. Otherewise the default sampling rate is used.

source

Sound(file)

Load a specified file as a Sound object.

source

Sound(fn,len;asseconds=true,rate=samplerate(),offset=0s)

Creates monaural sound where fn(t) returns the amplitudes for a given Range of time points (in seconds as a Float64). The function fn(t) should return values ranging between -1 and 1 as an iterable object, and should be just as long as t.

If asseconds is false, fn(i) returns the amplitudes for a given Range of sample indices (rather than time points).

If offset is specified, return the section of the sound starting from offset (rather than starting from 0 seconds).

source

Sounds.tone — Function.

tone(freq,length;[rate=samplerate()],[phase=0])

Creates a pure tone of the given frequency and length (in seconds).

source

Sounds.noise — Function.

noise(length;[rate=samplerate()],[rng=RandomDevice()])

Creates a period of white noise of the given length (in seconds).

source

Sounds.silence — Function.

silence(length;[rate=samplerate()])

Creates period of silence of the given length (in seconds).

source

Sounds.harmonic_complex — Function.

harmonic_complex(f0,harmonics,amps,length,
                 [rate=samplerate()],[phases=zeros(length(harmonics))])

Creates a harmonic complex of the given length, with the specified harmonics at the given amplitudes. This implementation is somewhat superior to simply summing a number of pure tones generated using tone, because it avoids beating in the sound that may occur due floating point errors.

source

Sounds.irn — Function.

irn(n,freq,length;[g=1],[rate=samplerate()],[rng=Base.GLOBAL_RNG])

Creates an iterated ripple $y_n(t)$ for a noise $y_0(t)$ according to the following formula.

$y_n(t) = y_{n-1}(t) + g⋅y_{n-1}(t-1/freq)$

source

Sounds.inHz — Function.

inHz(quantity)

Translate a particular quantity (usually a frequency) to a value in Hz.

Example

inHz(1.0kHz)

1000.0 Hz

source

Sounds.inframes — Function.

inframes(quantity,rate)

Translate the given quantity (usually a time) to a sample index, given a particualr samplerate.

Example

inframes(1s,44100Hz)

44100

source

Sounds.inseconds — Function.

inseconds(quantity)

Translate a particular quantity (usually a time) to a value in seconds.

Example

inseconds(50.0ms)

0.05 s

source

Interface

Sounds.samplerate — Function.

samplerate()

Return the current default sampling rate in units of Hz when constructing sounds. Also throws a warning since relying on implicit default values can be dangerous. Defaults to 44100 Hz.

source

samplerate(::Array)

Yields the same result as samplerate().

source

samplerate(::Sound)

Report the sampling rate of the sound in units of Hz.

source

Sounds.set_default_samplerate! — Function.

set_default_samplerate!(rate)

Changes the sampling rate returned by samplerate().

source

Sounds.nchannels — Method.

nchannels(sound)

Return the number of channels (1 for mono, 2 for stereo) in this sound.

source

Sounds.nframes — Method.

nframes(sound)

Returns the number of samples in the sound.

The number of samples is not always the same as the length of the sound. Stereo sounds have a length of 2 x nframes(sound).

source

Sounds.duration — Function.

duration(x;rate=samplerate(x))

Returns the duration of the sound.

The rate keyword is only avaiable for generic array-like objects. It is an error to pass a different rate to this function for a Sound

source

Sounds.left — Function.

left(sound)

Extract the left channel of a sound. For monaural sounds, left and right return the same value.

source

Sounds.right — Function.

right(sound)

Extract the right channel of a sound. For monaural sounds, left and right return the same value.

source

Sounds.ismono — Function.

ismono(x)

True if the sound is monaural.

source

Sounds.isstereo — Function.

isstereo

True if the sound is stereo.

source

Sounds.asmono — Function.

asmono(x)

Returns a monaural version of a sound (whether it is stereo or monaural).

source

Sounds.asstereo — Function.

asstereo(x)

Returns a stereo version of a sound (wether it is stereo or monaural).

source

Sound Manipulation

Sounds.highpass — Function.

highpass([x],high,[order=5])

High-pass filter the sound at the specified frequency.

Filtering uses a butterworth filter of the given order.

With one positional argument this returns a function f(x) that calls highpass(x,high;order=order)

source

Sounds.lowpass — Function.

lowpass([x],low,[order=5])

Low-pass filter the sound at the specified frequency.

Filtering uses a butterworth filter of the given order.

With one positional argument this returns a function f(x) that calls lowpass(x,low;order=order)

source

Sounds.bandpass — Function.

bandpass([x],low,high;[order=5])

Band-pass filter the sound at the specified frequencies.

Filtering uses a butterworth filter of the given order.

With two positional arguments this returns a function f(x) that calls bandpass(x,low,hight;order=order)

source

Sounds.bandstop — Function.

bandstop([x],low,high,[order=5])

Band-stop filter of the sound at the specified frequencies.

Filtering uses a butterworth filter of the given order.

With two positional arguments this returns a function f(x) that calls bandstop(x,low,hight;order=order)

source

Sounds.ramp — Function.

ramp([sound],[length=5ms])

Applies a half cosine ramp to start and end of the sound.

Ramps prevent clicks at the start and end of sounds.

When there is no sound argument, this returns a function f(x) that ramp(x,length)

source

Sounds.rampon — Function.

rampon([sound],[length=5ms])

Applies a half consine ramp to start of the sound.

When passed no sound argument, this returns a function f(x) which calls rampon(x,length)

source

Sounds.rampoff — Function.

rampoff([sound],[length=5ms])

Applies a half consine ramp to the end of the sound.

When passed no sound argument, this returns a function f(x) which calls rampoff(x,length)

source

Sounds.fadeto — Function.

fadeto([a],[b],[transition=50ms])

A smooth transition from a to b, overlapping the end of one with the start of the other by transition.

When passed a single sound, a, this returns a function f(b) that calls fadeto(a,b,transition).

source

Sounds.amplify — Function.

amplify([x],ratio)

Amplify (positive) or attenuate (negative) the sound by a given ratio, typically specified in decibels (e.g. amplify(x,10dB)).

Note: you can also directly multiply by a factor, e.g. x * 10dB, which has the same effect as this function.

With one positional argument, ratio, this returns a function f(x) which calls amplify(x,ratio).

source

Sounds.normpower — Function.

normpower(x)

Normalize the sound so it has a power of 1.

source

Sounds.mix — Function.

mix(x,...)

Add several sounds together so that they play at the same time.

Unlike normal addition, this acts as if each sound is padded with zeros at the end so that the lengths of all sounds match.

With one argument x, this returns a function f(y) that calls mix(x,y)

source

Sounds.envelope — Function.

envelope([x],y)

Mutliply several sounds together. Typically used to apply an amplitude envelope x to sound y.

Unlike normal multiplication, this acts as if each sound is padded with ones at the end so that the lengths of all sounds match.

With one argument x, this returns a function f(y) that calls envelope(x,y)

source

Sounds.dc_offset — Function.

dc_offset(length;[rate=44100Hz])

Creates a DC offset of unit value.

In other words, this just returns samples all with the value 1: silence is to zeros just as dc_offset is to ones.

Though this technically constructs a sound it is normally only used in combination with envelope, because it produces no audible sound. For example, it could be used to transition from a unmodulated to an amplitude modulated noise:

env = dc_offset(2s) |> fadeto(tone(8Hz,2s)) |> ramp(250ms)
sound = noise(duration(env)) |> envelope(env) |> normpower

source

Sounds.leftright — Function.

leftright(left,right)

Create a stereo sound from two monaural sounds.

source

DSP.Filters.resample — Method.

resample(x::Sound,new_rate;warn=true)

Returns a new sound representing x at the given sampling rate (in Hertz).

If you reduce the sampling rate, you will loose all frequencies in x that are above new_rate/2. Reducing the sampling rate will produce a warning unless warn is false.

source